@@ -3,15 +3,14 @@ package main
 import (
 	"context"
 	"encoding/base64"
-	"flag"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 	"strings"
 
 	"github.com/joho/godotenv"
-	"github.com/urfave/cli/v2"
+	"github.com/urfave/cli/v3"
 	"github.com/ybbus/jsonrpc/v3"
 )
 
@@ -30,7 +29,7 @@ func getSpec(
 	resp, err := client.Do(req)
 	checkErr(err)
 	defer resp.Body.Close()
-	bodyBytes, _ := ioutil.ReadAll(resp.Body)
+	bodyBytes, _ := io.ReadAll(resp.Body)
 	return jsonDecodeBytes(bodyBytes)
 }
 
@@ -49,7 +48,7 @@ func getDefaultURL(coin string) string {
 	return "http://" + host + ":" + port
 }
 
-func runCommand(c *cli.Context) (*jsonrpc.RPCResponse, map[string]interface{}, error) {
+func runCommand(c *cli.Command, help bool) (*jsonrpc.RPCResponse, map[string]interface{}, error) {
 	args := c.Args()
 	wallet := c.String("wallet")
 	contract := c.String("contract")
@@ -75,7 +74,10 @@ func runCommand(c *cli.Context) (*jsonrpc.RPCResponse, map[string]interface{}, e
 		},
 	})
 	// some magic to make array with the last element being a dictionary with xpub in it
-	sl := args.Slice()[1:]
+	sl := []string{}
+	if !help {
+		sl = args.Slice()[1:]
+	}
 	var params []interface{}
 	keyParams := map[string]interface{}{
 		"xpub": map[string]interface{}{
@@ -105,7 +107,11 @@ func runCommand(c *cli.Context) (*jsonrpc.RPCResponse, map[string]interface{}, e
 	}
 	params = append(params, keyParams)
 	// call RPC method
-	result, err := rpcClient.Call(context.Background(), args.Get(0), params)
+	command := "help"
+	if !help {
+		command = args.Get(0)
+	}
+	result, err := rpcClient.Call(context.Background(), command, params)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -122,232 +128,231 @@ func main() {
 		// Host:  "https://api.bitcart.ai",
 	}
 	rootOptions.Load()
-	app := cli.NewApp()
-	app.Name = "bitcart-cli"
-	app.Version = Version
-	app.HideHelp = true
-	app.Usage = "Call RPC methods from console"
-	app.UsageText = "bitcart-cli method [args]"
-	app.EnableBashCompletion = true
-	app.Flags = []cli.Flag{
-		&cli.BoolFlag{
-			Name:    "help",
-			Aliases: []string{"h"},
-			Usage:   "show help",
+	app := &cli.Command{
+		Name:                  "bitcart-cli",
+		Version:               Version,
+		HideHelp:              true,
+		Usage:                 "Call RPC methods from console",
+		UsageText:             "bitcart-cli method [args]",
+		EnableShellCompletion: true,
+		Flags: []cli.Flag{
+			&cli.BoolFlag{
+				Name:    "help",
+				Aliases: []string{"h"},
+				Usage:   "show help",
+			},
+			&cli.StringFlag{
+				Name:     "wallet",
+				Aliases:  []string{"w"},
+				Usage:    "specify wallet",
+				Required: false,
+				Sources:  cli.EnvVars("BITCART_WALLET"),
+			},
+			&cli.StringFlag{
+				Name:     "contract",
+				Usage:    "specify contract",
+				Required: false,
+				Sources:  cli.EnvVars("BITCART_CONTRACT"),
+			},
+			&cli.StringFlag{
+				Name:     "address",
+				Usage:    "specify address (XMR-only)",
+				Required: false,
+				Sources:  cli.EnvVars("BITCART_ADDRESS"),
+			},
+			&cli.BoolFlag{
+				Name:    "diskless",
+				Aliases: []string{"d"},
+				Usage:   "Load wallet in memory only",
+				Value:   false,
+				Sources: cli.EnvVars("BITCART_DISKLESS"),
+			},
+			&cli.StringFlag{
+				Name:    "coin",
+				Aliases: []string{"c"},
+				Usage:   "specify coin to use",
+				Value:   "btc",
+				Sources: cli.EnvVars("BITCART_COIN"),
+			},
+			&cli.StringFlag{
+				Name:    "user",
+				Aliases: []string{"u"},
+				Usage:   "specify daemon user",
+				Value:   "electrum",
+				Sources: cli.EnvVars("BITCART_LOGIN"),
+			},
+			&cli.StringFlag{
+				Name:    "password",
+				Aliases: []string{"p"},
+				Usage:   "specify daemon password",
+				Value:   "electrumz",
+				Sources: cli.EnvVars("BITCART_PASSWORD"),
+			},
+			&cli.StringFlag{
+				Name:     "url",
+				Aliases:  []string{"U"},
+				Usage:    "specify daemon URL (overrides defaults)",
+				Required: false,
+				Sources:  cli.EnvVars("BITCART_DAEMON_URL"),
+			},
+			&cli.BoolFlag{
+				Name:    "no-spec",
+				Usage:   "Disables spec fetching for better exceptions display",
+				Value:   false,
+				Sources: cli.EnvVars("BITCART_NO_SPEC"),
+			},
+			&cli.StringFlag{
+				Name:        "github-api",
+				Value:       "https://api.github.com",
+				Usage:       "Change the default endpoint to GitHub API for retrieving updates",
+				Destination: &rootOptions.GitHubAPI,
+			},
+			&cli.BoolFlag{
+				Name:        "skip-update-check",
+				Usage:       "Skip the check for updates check run before every command.",
+				Value:       skipUpdateByDefault(),
+				Destination: &rootOptions.SkipUpdateCheck,
+			},
 		},
-		&cli.StringFlag{
-			Name:     "wallet",
-			Aliases:  []string{"w"},
-			Usage:    "specify wallet",
-			Required: false,
-			EnvVars:  []string{"BITCART_WALLET"},
+		Before: func(ctx context.Context, cmd *cli.Command) (context.Context, error) {
+			if cmd.Args().Get(0) == "update" {
+				return ctx, nil
+			}
+			err := checkForUpdates(rootOptions)
+			if err != nil {
+				fmt.Printf("Error checking for updates: %s\n", err)
+			}
+			return ctx, nil
 		},
-		&cli.StringFlag{
-			Name:     "contract",
-			Usage:    "specify contract",
-			Required: false,
-			EnvVars:  []string{"BITCART_CONTRACT"},
-		},
-		&cli.StringFlag{
-			Name:     "address",
-			Usage:    "specify address (XMR-only)",
-			Required: false,
-			EnvVars:  []string{"BITCART_ADDRESS"},
-		},
-		&cli.BoolFlag{
-			Name:    "diskless",
-			Aliases: []string{"d"},
-			Usage:   "Load wallet in memory only",
-			Value:   false,
-			EnvVars: []string{"BITCART_DISKLESS"},
-		},
-		&cli.StringFlag{
-			Name:    "coin",
-			Aliases: []string{"c"},
-			Usage:   "specify coin to use",
-			Value:   "btc",
-			EnvVars: []string{"BITCART_COIN"},
-		},
-		&cli.StringFlag{
-			Name:    "user",
-			Aliases: []string{"u"},
-			Usage:   "specify daemon user",
-			Value:   "electrum",
-			EnvVars: []string{"BITCART_LOGIN"},
-		},
-		&cli.StringFlag{
-			Name:    "password",
-			Aliases: []string{"p"},
-			Usage:   "specify daemon password",
-			Value:   "electrumz",
-			EnvVars: []string{"BITCART_PASSWORD"},
-		},
-		&cli.StringFlag{
-			Name:     "url",
-			Aliases:  []string{"U"},
-			Usage:    "specify daemon URL (overrides defaults)",
-			Required: false,
-			EnvVars:  []string{"BITCART_DAEMON_URL"},
-		},
-		&cli.BoolFlag{
-			Name:    "no-spec",
-			Usage:   "Disables spec fetching for better exceptions display",
-			Value:   false,
-			EnvVars: []string{"BITCART_NO_SPEC"},
-		},
-		&cli.StringFlag{
-			Name:        "github-api",
-			Value:       "https://api.github.com",
-			Usage:       "Change the default endpoint to GitHub API for retrieving updates",
-			Destination: &rootOptions.GitHubAPI,
-		},
-		&cli.BoolFlag{
-			Name:        "skip-update-check",
-			Usage:       "Skip the check for updates check run before every command.",
-			Value:       skipUpdateByDefault(),
-			Destination: &rootOptions.SkipUpdateCheck,
-		},
-	}
-	app.Before = func(c *cli.Context) error {
-		if c.Args().Get(0) == "update" {
-			return nil
-		}
-		err := checkForUpdates(rootOptions)
-		if err != nil {
-			fmt.Printf("Error checking for updates: %s\n", err)
-		}
-		return nil
-	}
-	app.BashComplete = func(c *cli.Context) {
-		set := flag.NewFlagSet("app", 0)
-		checkErr(set.Parse([]string{"help"}))
-		output, _, err := runCommand(cli.NewContext(app, set, c))
-		if err != nil || output.Error != nil {
-			fmt.Println("plugin")
+		ShellComplete: func(ctx context.Context, cmd *cli.Command) {
+			output, _, err := runCommand(cmd, true)
+			if err != nil || output.Error != nil {
+				fmt.Println("plugin")
+				if updatesEnabled() {
+					fmt.Println("update")
+				}
+				return
+			}
+			output.Result = append(output.Result.([]interface{}), "plugin")
 			if updatesEnabled() {
-				fmt.Println("update")
+				output.Result = append(output.Result.([]interface{}), "update")
 			}
-			return
-		}
-		output.Result = append(output.Result.([]interface{}), "plugin")
-		if updatesEnabled() {
-			output.Result = append(output.Result.([]interface{}), "update")
-		}
-		for _, v := range output.Result.([]interface{}) {
-			fmt.Println(v)
-		}
-	}
-	app.Action = func(c *cli.Context) error {
-		args := c.Args()
-		if args.Len() >= 1 {
-			result, spec, err := runCommand(c)
-			checkErr(err)
-			// Print either error if found or result
-			if result.Error != nil {
-				if len(spec) != 0 {
-					if spec["error"] != nil {
-						exitErr(jsonEncode(spec["error"]))
+			for _, v := range output.Result.([]interface{}) {
+				fmt.Println(v)
+			}
+		},
+		Action: func(ctx context.Context, cmd *cli.Command) error {
+			args := cmd.Args()
+			if args.Len() >= 1 {
+				result, spec, err := runCommand(cmd, false)
+				checkErr(err)
+				// Print either error if found or result
+				if result.Error != nil {
+					if len(spec) != 0 {
+						if spec["error"] != nil {
+							exitErr(jsonEncode(spec["error"]))
+						}
+						exceptions := spec["exceptions"].(map[string]interface{})
+						errorCode := fmt.Sprint(result.Error.Code)
+						if exception, ok := exceptions[errorCode]; ok {
+							exception, _ := exception.(map[string]interface{})
+							exitErr(
+								exception["exc_name"].(string) + ": " + exception["docstring"].(string),
+							)
+						}
 					}
-					exceptions := spec["exceptions"].(map[string]interface{})
-					errorCode := fmt.Sprint(result.Error.Code)
-					if exception, ok := exceptions[errorCode]; ok {
-						exception, _ := exception.(map[string]interface{})
-						exitErr(
-							exception["exc_name"].(string) + ": " + exception["docstring"].(string),
-						)
-					}
-				}
-				exitErr(jsonEncode(result.Error))
-			} else {
-				var v, ok = result.Result.(string)
-				if ok {
-					smartPrint(v)
+					exitErr(jsonEncode(result.Error))
 				} else {
-					smartPrint(jsonEncode(result.Result))
+					var v, ok = result.Result.(string)
+					if ok {
+						smartPrint(v)
+					} else {
+						smartPrint(jsonEncode(result.Result))
+					}
+					return nil
 				}
-				return nil
+			} else {
+				checkErr(cli.ShowAppHelp(cmd))
 			}
-		} else {
-			checkErr(cli.ShowAppHelp(c))
-		}
-		return nil
-	}
-	app.Commands = []*cli.Command{
-		{
-			Name:  "plugin",
-			Usage: "Manage plugins",
-			Subcommands: []*cli.Command{
-				{
-					Name:      "init",
-					Action:    initPlugin,
-					Usage:     "Create a new plugin",
-					UsageText: "bitcart-cli plugin init <path>",
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:    "save",
-							Aliases: []string{"s"},
-							Usage:   "Save repository directories to not ask later",
-							Value:   false,
+			return nil
+		},
+		Commands: []*cli.Command{
+			{
+				Name:  "plugin",
+				Usage: "Manage plugins",
+				Commands: []*cli.Command{
+					{
+						Name:      "init",
+						Action:    initPlugin,
+						Usage:     "Create a new plugin",
+						UsageText: "bitcart-cli plugin init <path>",
+						Flags: []cli.Flag{
+							&cli.BoolFlag{
+								Name:    "save",
+								Aliases: []string{"s"},
+								Usage:   "Save repository directories to not ask later",
+								Value:   false,
+							},
 						},
 					},
-				},
-				{
-					Name:      "install",
-					Action:    installPlugin,
-					Usage:     "Install a plugin",
-					UsageText: "bitcart-cli plugin install [command options] <path>",
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:    "dev",
-							Usage:   "Install in development mode (symlink instead of copying)",
-							Value:   false,
-							Aliases: []string{"D"},
-						},
-						&cli.BoolFlag{
-							Name:    "save",
-							Aliases: []string{"s"},
-							Usage:   "Save repository directories to not ask later",
-							Value:   false,
-						},
-					},
-				},
-				{
-					Name:      "uninstall",
-					Action:    uninstallPlugin,
-					Usage:     "Uninstall a plugin",
-					UsageText: "bitcart-cli plugin uninstall <path>",
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:    "save",
-							Aliases: []string{"s"},
-							Usage:   "Save repository directories to not ask later",
-							Value:   false,
+					{
+						Name:      "install",
+						Action:    installPlugin,
+						Usage:     "Install a plugin",
+						UsageText: "bitcart-cli plugin install [command options] <path>",
+						Flags: []cli.Flag{
+							&cli.BoolFlag{
+								Name:    "dev",
+								Usage:   "Install in development mode (symlink instead of copying)",
+								Value:   false,
+								Aliases: []string{"D"},
+							},
+							&cli.BoolFlag{
+								Name:    "save",
+								Aliases: []string{"s"},
+								Usage:   "Save repository directories to not ask later",
+								Value:   false,
+							},
 						},
 					},
-				},
-				{
-					Name:      "validate",
-					Action:    validatePlugin,
-					Usage:     "Validate plugin manifest and common checks",
-					UsageText: "bitcart-cli plugin validate <path>",
-					Flags: []cli.Flag{
-						&cli.StringFlag{
-							Name:  "schema",
-							Usage: "Supply custom schema URL",
-							Value: schemaURL,
+					{
+						Name:      "uninstall",
+						Action:    uninstallPlugin,
+						Usage:     "Uninstall a plugin",
+						UsageText: "bitcart-cli plugin uninstall <path>",
+						Flags: []cli.Flag{
+							&cli.BoolFlag{
+								Name:    "save",
+								Aliases: []string{"s"},
+								Usage:   "Save repository directories to not ask later",
+								Value:   false,
+							},
 						},
 					},
-				},
-				{
-					Name:      "package",
-					Action:    packagePlugin,
-					Usage:     "Package plugin from its directory",
-					UsageText: "bitcart-cli plugin package [command options] <path>",
-					Flags: []cli.Flag{
-						&cli.BoolFlag{
-							Name:  "no-strip",
-							Usage: "Don't strip unneccesary files from the package (i.e. node_modules)",
-							Value: false,
+					{
+						Name:      "validate",
+						Action:    validatePlugin,
+						Usage:     "Validate plugin manifest and common checks",
+						UsageText: "bitcart-cli plugin validate <path>",
+						Flags: []cli.Flag{
+							&cli.StringFlag{
+								Name:  "schema",
+								Usage: "Supply custom schema URL",
+								Value: schemaURL,
+							},
+						},
+					},
+					{
+						Name:      "package",
+						Action:    packagePlugin,
+						Usage:     "Package plugin from its directory",
+						UsageText: "bitcart-cli plugin package [command options] <path>",
+						Flags: []cli.Flag{
+							&cli.BoolFlag{
+								Name:  "no-strip",
+								Usage: "Don't strip unneccesary files from the package (i.e. node_modules)",
+								Value: false,
+							},
 						},
 					},
 				},
@@ -358,7 +363,7 @@ func main() {
 		app.Commands = append(app.Commands, &cli.Command{
 			Name:  "update",
 			Usage: "CLI update operations",
-			Subcommands: []*cli.Command{
+			Commands: []*cli.Command{
 				{
 					Name:   "check",
 					Action: updateCLI,
@@ -373,5 +378,5 @@ func main() {
 		})
 	}
 	godotenv.Load(envFile) // nolint:errcheck
-	checkErr(app.Run(os.Args))
+	checkErr(app.Run(context.Background(), os.Args))
 }
